Creating BruTile.Desktop.DbCache project Grouping projects in solution
