+++ Tests/BruTile.Desktop.Tests/Utilities/Paths.cs
@@ -10,10 +10,15 @@ namespace BruTile.Tests.Utilities
         {
             get
             {
-                var codeBase = Assembly.GetExecutingAssembly().CodeBase;
+#if NET45
+                var asm = typeof(Paths).GetTypeInfo().Assembly;
+                var codeBase = asm.CodeBase;
                 var uri = new UriBuilder(codeBase);
                 var path = Uri.UnescapeDataString(uri.Path);
                 return Path.GetDirectoryName(path);
+#else
+                return AppContext.BaseDirectory;
+#endif
             }
         }
     }
